+++ src/Smartstore.Web/Areas/Admin/Models/Catalog/DeletedProductListModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Smartstore.Admin.Models.Catalog
+{
+    [LocalizedDisplay("Admin.Catalog.Products.List.")]
+    public class DeletedProductListModel : ProductListModel
+    {
+        [LocalizedDisplay("*SearchWithOrders")]
+        public bool? SearchWithOrders { get; set; }
+    }
+}