+++ changelog.md
@@ -57,6 +57,7 @@
   - #1262 Added OpenAI `o1` and `03-mini` to supported models.
   - Added a setting for the maximum number of completion tokens.
 - #1236 PayPal: Added setting to disable address application in case the store's address validation settings need to be respected.
+- #1274 Recycle Bin: Possibility to filter products with/without assignment to orders.
 
 ### Bugfixes